feat: reducers logic merged
