+++ src/store/dialogsReducer.js
@@ -1,20 +1,28 @@
 import { UPDATE_NEW_MESSAGE_BODY, SEND_MESSAGE } from './actionTypes';
 
-export const dialogsReducer = (state, action) => {
+const dialogsReducer = (state, action) => {
+  switch (action.type) {
+    case UPDATE_NEW_MESSAGE_BODY:
+      state.newMessageBody = action.body;
+      return state;
 
-    switch (action.type) {
-        case UPDATE_NEW_MESSAGE_BODY:
-            state.newMessageBody = action.body;
-            return state;
-            
-        case SEND_MESSAGE:
-            const body = state.newMessageBody;
+    case SEND_MESSAGE:
+      const body = state.newMessageBody;
 
-            state.newMessageBody = '';
-            state.messages.push({ id: 8, message: body });
-            return state;
+      state.newMessageBody = '';
+      state.messages.push({ id: 8, message: body });
+      return state;
 
-        default:
-            return state;
-    };
+    default:
+      return state;
+  }
 };
+
+const sendMessageCreator = () => ({ type: SEND_MESSAGE });
+
+const updateNewMessageBodyCreator = (body) => ({
+  type: UPDATE_NEW_MESSAGE_BODY,
+  body,
+});
+
+export { sendMessageCreator, updateNewMessageBodyCreator, dialogsReducer };
